refactor(ui): Isolate Mesop page definitions and standardize page structure
This commit refactors the application to improve modularity and enforce a standard pattern for creating Mesop pages.

Key changes include:

- All `@me.page` decorators have been moved from `main.py` into their respective page files within the `pages/` directory. This makes each page a self-contained component.

- Page-specific state classes (e.g., for the VTO page) have been co-located within their corresponding page file, and the redundant files in `state/` have been removed.

- The page layout has been standardized. All pages now correctly use the `page_scaffold` and `page_frame` components to ensure a consistent layout with proper padding and side navigation.

- `main.py` has been significantly cleaned up, now only containing FastAPI setup and module imports for Mesop's page discovery.

- The `developers_guide.md` has been updated to reflect this new, standardized process for creating pages.
