+++ experiments/veo-app/main.py
@@ -34,20 +34,21 @@ import pages.shop_the_look
 from app_factory import app
 from components.page_scaffold import page_scaffold
 from config import default as config
-from pages.about import about_page_content
-from pages.character_consistency import character_consistency_page_content
 from pages import chirp_3hd as chirp_3hd_page
 from pages import config as config_page
-from pages.edit_images import content as edit_images_content
-from pages.gemini_image_generation import gemini_image_gen_page_content
-from pages.gemini_tts import page as gemini_tts_page
-from pages.home import home_page_content
 from pages import imagen as imagen_page
 from pages import library as library_page
 from pages import lyria as lyria_page
 from pages import portraits as motion_portraits
 from pages import recontextualize as recontextualize_page
 from pages import starter_pack as starter_pack_page
+from pages import veo
+from pages import about as about_page
+from pages import character_consistency as character_consistency_page
+from pages.edit_images import content as edit_images_content
+from pages import gemini_image_generation as gemini_image_generation_page
+from pages import gemini_tts as gemini_tts_page
+from pages import home as home_page
 from pages.test_character_consistency import page as test_character_consistency_page
 from pages.test_index import page as test_index_page
 from pages.test_infinite_scroll import test_infinite_scroll_page
@@ -55,8 +56,7 @@ from pages.test_pixie_compositor import test_pixie_compositor_page
 from pages.test_uploader import test_uploader_page
 from pages.test_vto_prompt_generator import page as test_vto_prompt_generator_page
 from pages.test_worsfold_encoder import test_worsfold_encoder_page
-from pages import veo
-from pages.vto import vto
+from pages import vto as vto_page
 from state.state import AppState
 
 
@@ -160,56 +160,43 @@ async def set_request_context(request: Request, call_next):
     )
     return response
 
-@me.page(
-    path="/home",
-    title="GenMedia Creative Studio - v.next",
-    security_policy=me.SecurityPolicy(
-        dangerously_disable_trusted_types=True,
-    ),
-    stylesheets=[
-        "https://fonts.googleapis.com/css2?family=Google+Symbols:opsz,wght,FILL,GRAD,ROND@20..48,100..700,0..1,-50..200,0..100&icon_names=spark",
-    ]
-)
-def home_page():
-    """Main Page."""
-    state = me.state(AppState)
-    with page_scaffold(page_name="home"):  # pylint: disable=not-context-manager
-        home_page_content(state)
 
-@me.page(path="/gemini-tts", title="GenMedia Creative Studio - Gemini TTS")
-def gemini_tts_route():
-    with page_scaffold(page_name="gemini-tts"):  # pylint: disable=E1129:not-context-manager
-        gemini_tts_page()
 
-@me.page(path="/vto", title="GenMedia Creative Studio - Virtual Try-On")
-def vto_page():
-    with page_scaffold(page_name="vto"):  # pylint: disable=E1129:not-context-manager
-        vto()
 
-@me.page(path="/character_consistency", title="GenMedia Creative Studio - Character Consistency")
-def character_consistency_page():
-    with page_scaffold(page_name="character_consistency"):  # pylint: disable=E1129:not-context-manager
-        character_consistency_page_content()
 
-@me.page(path="/about", title="About - GenMedia Creative Studio")
-def about_page():
-    with page_scaffold(page_name="about"):  # pylint: disable=E1129:not-context-manager
-        about_page_content()
 
-@me.page(path="/gemini_image_generation", title="Gemini Image Generation - GenMedia Creative Studio")
-def gemini_image_generation_page():
-    with page_scaffold(page_name="gemini_image_generation"):  # pylint: disable=E1129:not-context-manager
-        gemini_image_gen_page_content()
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 # Test page routes are left as is, they don't need the scaffold
-me.page(path="/test_character_consistency", title="Test Character Consistency")(test_character_consistency_page)
+me.page(path="/test_character_consistency", title="Test Character Consistency")(
+    test_character_consistency_page
+)
 me.page(path="/test_index", title="Test Index")(test_index_page)
-me.page(path="/test_infinite_scroll", title="Test Infinite Scroll")(test_infinite_scroll_page)
-me.page(path="/test_pixie_compositor", title="Test Pixie Compositor")(test_pixie_compositor_page)
+me.page(path="/test_infinite_scroll", title="Test Infinite Scroll")(
+    test_infinite_scroll_page
+)
+me.page(path="/test_pixie_compositor", title="Test Pixie Compositor")(
+    test_pixie_compositor_page
+)
 me.page(path="/test_uploader", title="Test Uploader")(test_uploader_page)
-me.page(path="/test_vto_prompt_generator", title="Test VTO Prompt Generator")(test_vto_prompt_generator_page)
-me.page(path="/test_worsfold_encoder", title="Test Worsfold Encoder")(test_worsfold_encoder_page)
+me.page(path="/test_vto_prompt_generator", title="Test VTO Prompt Generator")(
+    test_vto_prompt_generator_page
+)
+me.page(path="/test_worsfold_encoder", title="Test Worsfold Encoder")(
+    test_worsfold_encoder_page
+)
 
 
 @app.get("/")
@@ -267,4 +254,3 @@ if __name__ == "__main__":
         timeout_graceful_shutdown=0,
         proxy_headers=True,
     )
-
