+++ experiments/veo-app/pages/gemini_tts.py
@@ -68,7 +68,8 @@ def page():
     """Renders the Gemini TTS page."""
     state = me.state(GeminiTtsState)
 
-    with page_frame():  # pylint: disable=E1129
+    with page_scaffold(page_name="gemini-tts"):
+        with page_frame():
             header(
                 "Gemini Text-to-Speech",
                 "record_voice_over",
